requirements.txt
alembic==1.13.2
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
customtkinter==5.2.2
darkdetect==0.8.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.5
MarkupSafe==2.1.5
multipart==0.2.5
orjson==3.9.15
packaging==24.1
passlib==1.7.4
pillow==10.4.0
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
Python_IO==0.3
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.30
starlette==0.36.3
typing_extensions==4.10.0
ujson==5.9.0
urllib3==2.2.2
uvicorn==0.28.0
watchfiles==0.21.0
websockets==12.0
